rework the csp section
the worker pools were implemented wrong, and the code was adjusted as such.

originally i was having problems on how to signal each worker pool to close
the channel, because we don't know the amount of items that there will be
in each queue.

from the go by example sample code, it seems that they rely on one fact that
thet can guarantee to be true: the initial amount of work.

if you can keep the queue sizes to be the same (by sending false to the channel
for example, just so it takes up space) you can just finally iterate over the
same amount of items towards the end and you'll be sure you're draining the
channels and quitting gracefully.

not sure if this is idiomatic, but this seems to work.